+++ src/core/queue/order/CHANGELOG.md
@@ -9,6 +9,14 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.??.?? (2021-??-??)
+
+#### :boom: Breaking Change
+
+* Renamed `Tasks` to `InnerQueue`
+* Renamed `CreateTasks` to `CreateInnerQueue`
+* Renamed `QueueOptions.tasksFactory` to `QueueOptions.queueFactory`
+
 ## v3.20.0 (2020-07-05)
 
 #### :rocket: New Feature